Add `minimist` package (matching version from `zx`) since it's not yet exported
Despite the `zx` docs suggesting that `minimist` is exported from `zx`,
that's not actually true. The code has been pushed for this feature, but
it's not released to `npm` yet.

Accordingly, manually install `minimist` so that we can import from it.
Be careful to match the exact version (`1.2.8`) which was resolved when
installing it as a dependency of `zx` (discovered through `yarn info -R minimist`
and then added with `yarn add -E minimist@1.2.8`).

See: https://github.com/google/zx/pull/661