Add scripts for prettifying, and for downloading top 100, extensions
Prettify extension source code with a normal-ish prettier config,
except that line length is set to 140 characters (since extension
bundles tend to include a lot of webpacked code, which is highly
nested and is counterintuitively more readable with longer line lengths).
Ignore any prettier errors (TODO: think about whether we want to save
the Prettier logs, as we probably do - any Prettier errors could potentially
indicate a bug due to bad syntax, etc.)

Add script for generating the top extensions file, by parsing one
that is expected to be at `extensions-2021.json` (but this is a ~50mb file
so we don't check it into the repository, and we make sure to tell TypeScript
to ignore it).

Add a script for downloading all of the top 100 extensions, and prettifying
the source code of each of them.

Also do some refactoring.