Allow fetching 1,000 updates at once, by reducing payload size
Previously it was not possible to send an updatecheck fro 1,000
extensons at once, because the request body would be too large. But
it was possible to remove some redundant fields in the body so that
we can in fact send 1,000 (and maybe more, who knows) extensions in
one request.

(Update: trying to send 1,500 causes the response to return only 500)